[desktop] Fixes edge case where getPreviewFile() would return null.
When the media torrent transfer was created by the user running
FrostWire the logic in getPreviewFile() would fail as it would
only consider downloaded files, meaning downloaded bytes was always zero.

This fix might not be ideal and it could break if the sharing ratio
string returned by libtorrent changes from "Infinity" to something else.

Maybe there's a more reliable way to check if the preview file is on disk,
complete (and non-altered, maybe we can check once for the chunk hashes)

For now this works well.