[desktop] Revert implementation of StreamableSearchResults on UISearchResults.
I hadn't noticed that UISearchResults had a .getSearchResult() method.
The bug was on SearchResultActionsRenderer.isStreamableSourceBeingPlayed()
which was asking if the UISearchResult passed was a StreamableSearchResult,
this made me think the API had changed and we had forgotten to make those
implement StreamableSearchResult, when all it had to do was ask its delegate.