[desktop] When transfers are started we make sure ALL transfers are shown.
I attempted to show transfers on their downloading or seeding state depending
on what I thought they should be, but this automatic picking of the filter
is a terrible experience as transfers can change state rapidly and then the user
may think they didn't donwload or they're gone.

Instead, whenever we depend on an automatic switch to the transfers tab it's better
to leave the experience as it used to be, show all. If the user wants to filter transfers
this must be a voluntary action. The experience is much nicer.