[android] Dismisses shutdown/back-home dialogs before things happen.
Since these dialogs are handled with a Dialog listening interface and
the onDialogClick can't provide a reference to the source dialog, these
YesNoDialogs can now be provided with a flag to dismiss the dialog before
dispatching the listener's onDialogClick.

Now when the user is about to leave, the dialog doesn't appear stuck and the
experience is more responsive.