[android] Avoid IndexOutOfBoundsException on BrowsePeerFragment.onLoadFinished()
If the BrowsePeerFragment manager *for some reason* returns a null list or an
empty list for a given file type, the FileListAdapter's visualList will have
0 elements.

When AbstractListAdapter.getItem() is triggered it crashes with an IndexOutOfBoundsException.

This might be happening between refreshes, as getItem() is often seen asking high indices
when the underlying visualList has 0 elements.