[android] Fixes bug where app wouldn't shutdown.
The isShutdown() method was always failing because it used `getIntent()`.
Added a parameter to it and invoked it on the `MainActivity.onNewIntent()`
method.

My belief is that this bug surfaced when the MainActivity was converted
to `launchMode=singleTask` https://github.com/frostwire/frostwire/commit/b2e211954450f128af553ca5a88ef158081c83d8#diff-9832a7bd42037dd9b6595bb2c5eea58bR63

When new Intents were fired on it, it wouldn't call `onCreate()`, it would go straight to
`onResume()`, and even though I also tried putting the logic to check the intent there,
it wouldn't work because it would be working with the original intent, not the one that
was just sent.