[android] test, don't go to idle when stopping.
seems to stop faster and finish the audio player activity faster this way. Trying to understand why the player keeps coming back even after we've stopped it completely, might be something else entirely that brings the last song back. Trying to see if this is what's making the app restart by itself, perhaps it's not the player, perhaps it's the new MobFox network that keeps something in the background.