[android] Fixes auto-restarting on exit issue.
the issue was MusicPlaybackService's shutdown logic.

The way this service shuts itself down is by receiving a delayed PendingIntent (this is a way to start a service at a later time even if the app isn't running) via Android's AlarmManager. This shutdown was being scheduled in 60 seconds.

If the service is started normally with Context.startService() it's onStartCommand() method returns START_STICKY, but if it's started with the special SHUTDOWN command, it checks if there's anything playing, and if it's not playing, it releases everything and returns START_NOT_STICKY.