[desktop] Fixes false positive VPN detection issue
After VPN technology updates, it's no longer reliable to query for network interfaces that contain the string 'tun' in them.
Nowadays such interfaces are added and active even if they're not part of the ip route table.
We're using a similar technique for posix to detect if the tunnel interface is the parting ip on the route table, every 20 seconds