[desktop] loads lyrics asynchrounously in playlist item
When a single song is selected from a playlist a tag parser is launched in a background executor.
When done the fetched lyric (or empty if nothing found) is passed as a reference inside a
runnable that will be executed in the UI thread for whatever purposes, in our case we'll display the song
lyric in the playlist screen.