[android] Fixes file opening issues for default storage
It seems that after SDK 24 we cannot pass a file:// URI to another app via Intent data
since all the permissions to interact with such file are a responsability of our app
android now decided we need to have a content provider for the files we send to other intents

This is working so far for our default save location, but it's problematic when the user
changes the default save location to say an external storage device since the FileProvider
documentation says that we must provide the directories before-hand when we declare
the provider in the manifest.

it is my hope that we'll find a way to manipulate such paths programatically somehow. Perhaps
you've read about this already @aldenml