[android] Fixes issue where My Files SearchView was being hidden right after being shown.
The culprit was a runnable sent from the ToolbarActionBar which would invoke populateOptionsMenu. This runnable was sent whenever the activity's invalidateOptionsMenu() method was called. After a search, the TransfersFragment.onTime() method was found to have a periodic call to invalidateOptionsMenu() from a recent fix on the transfer's menu. The fix was making sure such invalidation only happened when the TransfersFragment was visible. 14 characters.  The rest of the code changes had to do with a hard coded string and tighter checks on when to perform the filtering