[android] anti-ANR MusicPlaybackService refactors
Addresses tickets #532 and #533.

Operations like .openXXX() and .duration() may take a long time and cause
ANRs if invoked in the main thread. The code has been refactored to use callbacks
to the main thread and avoid the numerous ANR reports we've been getting in the google play console

Also fixes a possible NPE on ensurePlayListCapacity