[android] MainActivity.hideFragments() - Issue #556
don't crash if weaker transaction commit fails. Looking at Android's souce code it seems the IllegalStateException will occur if the BackStackRecord.mCommitted property has already been set to true. This makes me wonder if this happens because this hideFragments() call is being called concurrently as the real source of the bug