[android] RelativeLayout positioning fix
The ScrollView applied to TransfersNoSeedsView so that a user can see a full message on horizontal views was placed last in the RelativeLayout, which means it displayed first covering other views. Because of that a user would not be able to click on any download to see the context menu or see its files as the scroll view was on top.