[android] issue #598 fix
Minimize the number of calls to the MusicService on the main thread.
It was aggresively calling all the way down to the music service every 500ms to update the
user interface. In many ocassions these calls would take a long time to come back thus causing
thousands of ANR reports.

Along the way the Debug context leak detection tool was fixed helping sort out 2 false positives.

This implementation minimizes the number of static Runnables submited to just 3. I'm not certain if we should be
using the same thread pool for all these things. Perhaps the Engine's getThreadPool() could use a parameter and
internally work with multiple pools for distinct purposes. This requires a design conversation if we move forward
in that direction, so that we can group threads accordingly in the minimum number of thread pools as possible.