[android] SearchHeader banner reload bug fix
Did not have to be so aggresive about destroying the mopub view. When destroyed, it gets rid of its internal controller.
A mopub view that has a null internal controller aborts its ad loading logic.
That controller can only be created when the mopub view is instantiated, this only happened onCreate(), therefore only one ad was displayed.