[android] finally resolved rotation issue
It seems the pager adapter and the programatic approach to adding transfers ended up in fragments not being able to be tracked correctly by their tags when the activity was reloaded.

Some of the fragments we ended up with inside detailFragments[] wouldn't be added, however, we luckily have methods in the parent AbstractFragment class to keep track of fragments that have been added (since Android only provides these after API 21 or so), by looking at the added fragments I'm able to find the correct instance and replace it inside our internal detailFragments array.

Writing this commit message makes me want to use that list, and get rid of the internal detailFragments array, to keep track of the fragments used, they could easily be filtered if they're instances of AbstractTransferDetailFragment