[android] MainApplication starts BTEngine in individual thread
When simulating a slow BTEngineInitializer taking up to a minute to start and running this Runnable on the Engine Blocking Thread Pool,
it would make the Thread Pool queue up more than 20 tasks very quickly, which in production I believe would lead to OOM errors.

After moving this very neural initialization to a thread of its own the issue no longer occurs since other Runnables in the threadpool
are free to continue (Except for those waiting for the context latch)

I believe that such Runnables that make use of BTEngine.instance() calls should only make such use if the application cannot function without them, for example,
when registering setting listeners and other startup/bootstrap related tasks.

If a Runnable makes use of BTEngine.instance() and it's not a life/death use of BTEngine, it should end its thread if it encounters that BTEngine hasn't been initialized yet (ctx == null)