[android] No more Runtime error on ConfigurationManager.create()
ConfigurationManager.instance() may have been called while ConfigurationManager.create() was still busy.Added an AtomicReference<State> to synchronize the creation of the singleton and wait for it if necessary with a countdown latch (only once, if this situation ever occured).

The IO portion of the creation of the configuration manager is forced to a background thread.

We also ensure every editor.apply() background call does not occur in the main thread.

Even though SharedPreferencesImpl is supposed to make the .apply() call in a background thread, there's a countdown latch that is blocking the main thread.

We make sure our class is not guilty of making this happen. I have to check if Apollo uses a preference editor on the main thread.