[android] Cancel All Notifications refactor on EngineService
When starting the app the main thread could be held gettings the notification system service, which in android 7 has the possibility of a security exception (catched already), then all notifications are asked to be cancelled.

This code was repeated also when shutting down, also on the main thread

This has now been moved to a background runnable to make startup and shutdown faster to the user.