[android] commented custom ImageCache usage to allow migration to latest picasso 3.0.0-SNAPSHOT
The crude fact is that current versions of picasso does not support in
any reasonable way any sort of custom cache, which previously allowed
for a very fine granularity of which bytes to cache. Now, we need to test
in production to actually see if in fact, it was that necessary. It's
not clear, looking at possible load patterns, if it really is. It's quite
possible that our customization using disk actually made the performance
worse.