[all] SearchManager creation refactor
Instead of specifying the number of threads on an internal class loader for the singleton, we now have a SearchManager.create(ExecutorService) that should be called only once to create the Future<SearchManager> singleton instance.

I've done this to experiment on this branch with different executors for different types of search, in order to have dedicated threads for specific search sources. This is also thinking about P2P search having its own Thread Pool.