[android] Prevents system error on restart
When the app was being shutdown a second time it was trying to unregister a receiver that hadn't been registered.
This was because engineServiceStarter wasn't being called upon restart.
engineServiceStarter was an instance method that was called upon onApplicationCreate.
This method needs to be called in the background because it touched the configuration manager at some point.
I could not invoke it with async(getApplication(), this.engineServiceStarter) because the executor verifier would complain about
possible leaks, therefore I had to make the method static, however, async does not take two context parameters and a method so I pass the references to engine (which I now need to perform the internal assigments it used to do) and the application in a weak reference holding class. Perhaps this can be simplified