[desktop] IP Filter download/import refactor
The previous method of having an InputStream that could be either a GZippedInputStream or PipedInputStream proved to be very unreliable when trying to import multiple times.

Now it's a simpler algorithm. Regardless of the source of the filter file, it's first gunzipped if necessary and then we do the actual import from a regular FileInputStream.