[android] com.android.tools.build:gradle 3.2.1 -> 3.4.0
Finally figured out why we couldn't get past 3.2.1.
The error that would always come up had to do with the Keystore not being able to be read or tampered
I figured the order of execution changed and even though we were reading the password using a Swing dialog
it was already too late.

Moved that code to the first signingOptions section of the gradle script and now we can build with the latest stable gradle API 3.4.0