[android] NotificationUpdateDemon TimerSubscription fix
When the application is exited having a static TimerSubscription instance avoids an issue on which on Exit, a new EngineService is created and in the process a new NotificationUpdateDemon, which has a new instance of TimerSubscription which we cannot unsubscribe. By keeping a single static instance, when we mark it as unsubscribed all the logic works. Also we re-use the instance to resubscribe.