[android] Fixes issue where an existing partial torrent transfer could not add more items from a .torrent in My files
it's not working 100% of the way, trying to figure out why the uiBittorrentDownload in the Transfer Detail fragment isn't updated all the times, it works most of the times.
Gets wonky if the transfer is paused sometimes, but if you try again it will add all of the specified files. Gotta come back to this