[android] continuation of 3b6e8383180696202c7589e960c7aeb5a9534b6d
Recovers if it loses reference to fragment manager. Simple, effective but possibly risky approach of keeping a static reference to the living SearchFragment in a static variable in case the weak reference is lost. The reference is always removed and rewritten when the search fragment is destroyed, created or resumed. Gotta keep an eye for OOM errors after this hack