[android] No more weird flickering in RecyclerViews in Transfer details
1. Instead of clearing the adapter items and adding all the new elements, we replace the items one by one and use the adapter.notifyItemChanged methods. If elements are added or removed (like in the case of Peers), we make sure the adapter's item list will grow or shrink as necessary.

Then we proceded to implement this logic in AbstractTransferDetailFragment as a static method

and used this method in the adapters for files and peers.

On both cases we pre-sorted both lists by a certain criteria, download progress for files, total download + total upload speeds in the case of peers, both in decreasing order so you can see what files are being downloaded first and what peers are helping the most.