[telluride] try shutting down instead with a simple os.kill - thanks @jjgalvez
It seems Flask has deprecated request.environ.get('werkzeug.server.shutdown'), which was a hack anyway.
Flask does not provide a shutdown function because it's really not meant to be run anywhere except a development environment.
Flask apps are meant to be embedded into WSGI servers, which should in turn provide ways to be shutdown.
Maybe it makes sense to send a patch to Flask to address this need. Sometimes you just need a simple Flask server without much performance and whatever else proper WSGI servers provide.

This works so far on macos for m1 processors, have to test on windows and linux