[desktop] SearchEngines are first separated into simple (single http request) and crawleable.
Then when we add them to the list of search engines to use when performing searches, at the start of the list we have all of the single request ones, so that when sent to the SearchManager thread pools, they are added to the single request pool first.

This can be done perfectly if we add a new property to the search performers that the search manager can use.

We'll leave such optimization for later when we have merged the android implementation.