[Desktop] Fix search bar clear button overlap
Commit 0a9f7fae5ada95544e4767ca36b3801b996a1f70 introduced proper theme
switching between the light and new dark themes. However, at the same
time, it also introduced a bug that caused the search icon not to show
up and the clear button to overlap. This was because if the textfield's
layout was not a BorderLayout, we would give it a new BorderLayout,
causing the layout to be reset (from what I understand) and thus causing
this issue.

To fix this issue, this commit simply reverts that change with a few
differences. I also reintroduce the try-catch block here, but we
explicitly catch a null pointer exception just in case our component,
for some reason, is null.

I'm not sure why the previous change was made. It seemed like it had no
relevance to the commit, but based on the comment, it sounds like there
was a good reason. Maybe we'll figure it out sooner than later.