Remove SystemUtils.openFile()
This method was only used on Windows Vista. Windows 7 and later used
ProcessBuilder, the same as all other platforms.

Since Windows Vista has been unsupported for a long while, I think it's
safe to get rid of this member function as it was only being used in one
place on an unsupported platform. By removing this method, this allows
us to further clean up the libSystemUtilities native code on Windows.