SystemUtilities: Clean up Windows code
This drops FullScreen.cpp and FullScreen.h completely as they went
unused. Additionally, this also drops a lot of unused code in Shell.cpp
and Shell.h too as those methods remained unused. Finally, this updates
the associated Visual Studio projects. I edited the
SystemUtilities.vcxproj file manually to remove FullScreen.cpp.
Hopefully this didn't break the project, but if needed I can update it
on Windows. I also dropped the other project file as it appears that
went unused. (The last every commit for that file was at the beginning
of this repo).