Drop libSystemUtilities support for macOS
Surprisingly, the actual libSystemUtilities support for macOS is just
one function, which also went unused. Therefore, we can go ahead and
drop all macOS-related things for SystemUtilities.

After this change, SystemUtilities effectively becomes a Windows-only
class. This means that the majority of work on SystemUtilities will now
be Windows-related. There are still some binary blobs required for
macOS, however, but more work on that will continue.