[desktop] libSystemUtilities minimization rebased (#1058)
* Remove FileUtils.setWriteable()

This method was effective obsolete due to the File.setWritable() method
that also did the same thing. In turn, this allows us to reduce our
dependency on SystemUtilities.

* Refactor trash handling

Use Desktop.moveToTrash() instead of platform-specific methods. This
further reduces on native methods of SystemUtilities, pushing us further
from moving away from native SystemUtilities functions in general and
therefore allowing us to drop the pre-compiled native binaries.

Note that moving files to the trash is still unsupported on Linux. This
is due to a limitation with Desktop.moveToTrash() apparently not being
supported on Linux. The JDK will have to update this method so Linux is
properly supported. Until then, only Windows and macOS are supported, but
now they use a common implementation rather than platform-specific
implementations.

* Linux: Remove SystemUtilities

The only Java method Linux's SystemUtilities was used for was unused,
therefore making all of this native code redundant. Therefore, drop all
the Linux-related source code and libraries associated with it as we
won't be using it anymore.

Signed-off-by: Avery King <gperson@disroot.org>

* Get rid of SystemUtilities.toggleFullScreen()

This method wasn't used by anything, so remove it along with its native
counterpart.

Signed-off-by: Avery King <gperson@disroot.org>

* Remove SystemUtils.openFile()

This method was only used on Windows Vista. Windows 7 and later used
ProcessBuilder, the same as all other platforms.

Since Windows Vista has been unsupported for a long while, I think it's
safe to get rid of this member function as it was only being used in one
place on an unsupported platform. By removing this method, this allows
us to further clean up the libSystemUtilities native code on Windows.

* SystemUtilities: Clean up Windows code

This drops FullScreen.cpp and FullScreen.h completely as they went
unused. Additionally, this also drops a lot of unused code in Shell.cpp
and Shell.h too as those methods remained unused. Finally, this updates
the associated Visual Studio projects. I edited the
SystemUtilities.vcxproj file manually to remove FullScreen.cpp.
Hopefully this didn't break the project, but if needed I can update it
on Windows. I also dropped the other project file as it appears that
went unused. (The last every commit for that file was at the beginning
of this repo).

* Drop libSystemUtilities support for macOS

Surprisingly, the actual libSystemUtilities support for macOS is just
one function, which also went unused. Therefore, we can go ahead and
drop all macOS-related things for SystemUtilities.

After this change, SystemUtilities effectively becomes a Windows-only
class. This means that the majority of work on SystemUtilities will now
be Windows-related. There are still some binary blobs required for
macOS, however, but more work on that will continue.

* [desktop] Fix Linux build

com.frostwire.jlibtorrent-linux64 doesn't exist, but
com.frostwire.jlibtorrent-linux does. Change it to the latter for now.
Not sure if the former is intended at some point, but for now, this
works.

---------

Signed-off-by: Avery King <gperson@disroot.org>