[desktop] Fix Welcome Window text color under the dark theme
When switching between FrostWire 6.13.2 and the latest master, if the
dark theme was previously set, the introductory text in the welcome
readable isn't very readable. While this may be an edge case, I don't
think fixing it really affects anything much, so it's been fixed.

Also, this can really help if we ever implement dynamic theming based on
the system's light/dark mode preference.