[desktop] Add GitHub Actions workflow
This opens the gateways to allowing not only for easier and more
transparent releases but also for nightly builds of FrostWire as well.
Right now, this allows us to ensure that FrostWire at least builds.

Note that we only have Ubuntu for now, but it's my intention to bring in
Windows and macOS runners into this workflow as well so we can verify
that FrostWire runs on all platforms as well.